Disable submitURL by default
Not all bots want to submit the test-suite results to perf, so
making the parameter null by default and requesting buildbots
to pass the URL at build time.

This will reduce the number of system failures given that the
perf system is rather unstable recently. In the long run it will
at least de-couple the failures from both systems.


git-svn-id: https://llvm.org/svn/llvm-project@204382 91177308-0d34-0410-b5e6-96231b3b80d8